Move center_out_reach task settings to its own module that can be imported elsewhere
